% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateFactors.R
\name{annotateFactors}
\alias{annotateFactors}
\alias{annotateFullDataset}
\alias{formatMetadata}
\title{Format sample data files as needed for sociodemographic and smoking analyses.}
\usage{
annotateFactors(NYC_HANES)

annotateFullDataset(dat)

formatMetadata(dat)
}
\arguments{
\item{NYC_HANES}{An object of class \code{pseq}.}

\item{dat}{Data.frame.}
}
\description{
Format sample data files as needed for sociodemographic and smoking analyses.
}
\details{
These are convenience functions that accomplish all recoding and formatting of sample data for consistency across sociodemographics and smoking analyses. 
\code{annotateFactors} formats the \code{phyloseq} object of the n=297 pilot study, and \code{annotateFullDataset} formats the full NYC HANES data.frame. \code{formatMetadata} produces a dataset where factor levels represent very nice-looking labels for sociodemographics variables used in plotting and tables.
}
