% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadQiimeData.R
\name{loadQiimeData}
\alias{loadQiimeData}
\title{Import NYC HANES-II microbiome data as a \code{phyloseq} object.}
\usage{
loadQiimeData()
}
\description{
Import NYC HANES-II microbiome data as a \code{phyloseq} object.
}
