% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_plots.R
\name{plot_abundance}
\alias{plot_abundance}
\alias{plot_abundance_by}
\title{Area and bar plots of sample-level relative abundance at specified taxonomic levels}
\usage{
plot_abundance(
  pseq,
  top_n = 8,
  taxrank = "Phylum",
  prop_of = "top_n",
  type = "area",
  orderby_OTU,
  PALETTE
)

plot_abundance_by(
  pseq,
  brewerpal = "Set1",
  by,
  type = c("area", "bar"),
  ttle = paste0("Relative phylum abundances by ", by)
)
}
\arguments{
\item{pseq}{An object of class \code{phyloseq}}

\item{top_n}{Integer. Number of taxonomic groups to display, sorted by relative abundance}

\item{taxrank}{Character. Taxonomic rank to display. Should be a column name of the \code{taxa_table} in \code{pseq}.}

\item{prop_of}{Character. Either "top_n" or "total". Indicates whether heights of area/bars should represent the proportion out of the displayed taxonomic groups, or the out of the total.}

\item{type}{Character. Either area or bar, specifying which type of plot. Area displays all samples left to right, whereas bars give mean abundances.}

\item{orderby_OTU}{Character (optional). The OTU name (i.e. rowname in \code{otu_table}) whose relative abundance you would like to sort by. This is what creates the visual gradient from left to right of, say, Firmicutes.}

\item{PALETTE}{Character of same length \code{top_n} of palette for area or bar chart.}

\item{brewerpal}{Character. Name of RColorBrewer palette to use.}

\item{by}{Character. Variable in \code{sample_data} to stratify plots by.}

\item{ttle}{Character. Plot title.}
}
\description{
Area and bar plots of sample-level relative abundance at specified taxonomic levels
}
