% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_plots.R
\name{plot_alpha_by}
\alias{plot_alpha_by}
\alias{plot_beta_by}
\title{Alpha and beta diversity plots stratified by sample variables}
\usage{
plot_alpha_by(pseq, vars, metadata = sample_data(pseq)[, vars],
  notch = TRUE)

plot_beta_by(pseq, vars, dist_mtrx)
}
\arguments{
\item{pseq}{An object of class \code{phyloseq}}

\item{vars}{Character vector. Names of variables in \code{sample_data} to stratify by.}

\item{metadata}{Data.frame. Defaults to \code{sample_data(pseq)[,vars]}, but this allows the use of say, a calculated variable without having to modify the \code{pseq}.}

\item{notch}{Logical. Whether or not to return notched boxplots.}

\item{dist_mtrx}{Matrix. Matrix of between-group distances, e.g. as returned by \code{phyloseq::distance()}.}
}
\description{
Alpha and beta diversity plots stratified by sample variables
}
