% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_dotplots.R
\name{plot_edgeR}
\alias{plot_edgeR}
\title{Dot plots based on edgeR results for microbiome data}
\usage{
plot_edgeR(formla, ttle = NULL, varname = NULL, pseq = NYC_HANES,
  coef = 2, printIfEmpty = FALSE, invisbl = FALSE, color = "FDR",
  sortby = NULL, ...)
}
\arguments{
\item{formla}{formula. specifies the design matrix used by \code{edgeR::glmFit}.}

\item{ttle}{character. Title of plot.}

\item{varname}{character. Label for independent variable, for plotting.}

\item{pseq}{object of class \code{phyloseq}}

\item{coef}{integer. Specifies which linear model coefficient to test (default 2).}

\item{printIfEmpty}{logical. If no results pass alpha, whether or not to plot an empty plot.}

\item{invisbl}{logical. Whether or not to forgoe plotting. Either way, invisibly returns the \code{topTags} object returned by \code{get_edgeR_results}.}

\item{color}{character. Variable  to color by, from either the \code{tax_table} or the \code{topTags$table}.}

\item{sortby}{character. Variable to sort the plot by, from either the \code{tax_table} or the \code{topTags$table}.}

\item{...}{further arguments passed to \code{get_edgeR_results}.}
}
\description{
Plots results of \code{get_edgeR_results_all_levels()} using the \code{ggplot2} engine.
}
