% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettytable_functions.R
\name{prettytable_autoindent}
\alias{prettytable_autoindent}
\title{Create an auto-indented table based on tableone::CreateTableOne.}
\usage{
prettytable_autoindent(vars, data, nspaces = 6, includeNA = TRUE,
  nonnormal = NULL, ...)
}
\arguments{
\item{vars}{Character vector. Names of variables to summarize}

\item{data}{Data.frame. Where to search for \code{vars}.}

\item{nspaces}{Integer. Number of spaces to add to the beginning of indented rows.}

\item{includeNA}{Logical. Whether or not to treat NAs as a factor level.}

\item{nonnormal}{Character vector. Names of variables in \code{vars} to treat as non-normally distributed for summary and test functions.}

\item{...}{Arguments passed to \code{CreateTableOne}}
}
\description{
Create an auto-indented table based on tableone::CreateTableOne.
}
