% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edgeR_results.R
\name{get_edgeR_results}
\alias{get_edgeR_results}
\alias{get_edger_results_all_levels}
\alias{edger_list_to_data.frame}
\alias{get_all_edgeR_models}
\title{User friendly interface to edgeR for microbiome data}
\usage{
get_edgeR_results(
  formla,
  pseq = NYC_HANES,
  method = c("BH", "IHW"),
  coef = 2,
  alph = 0.01,
  filtering = method[1] == "BH",
  countMinimum = 8,
  percentMinimumHaveCount = NULL,
  nMinimumHaveCount = 3,
  NA.RM = TRUE
)

get_edger_results_all_levels(formla, ...)

edger_list_to_data.frame(list_models)

get_all_edgeR_models(
  vars,
  varlabels = vars,
  adjusted_for,
  to.data.frame = TRUE,
  ...
)
}
\arguments{
\item{formla}{formula. specifies the design matrix used by \code{edgeR::glmFit}.}

\item{pseq}{object of class \code{phyloseq}}

\item{method}{character. Specifies the method of multiple testing correction to apply, either "BH" (Benjamini-Hochberg) or "IHW" (independent hypothesis weighting).}

\item{coef}{integer. Specifies which linear model coefficient to test (default 2).}

\item{alph}{numeric. Specifies what FDR level is considered alpha, and only keeps results with FDR less than this number.}

\item{filtering}{logical. Whether or not to apply pre-filtering.}

\item{countMinimum}{integer. If \code{filtering==TRUE}, the minimum count required for an OTU to be retained. Requires either \code{percentMinimumHaveCount} or \code{nMinimumHaveCount}.}

\item{percentMinimumHaveCount}{numeric. If \code{filtering==TRUE}, the minimum percentage of samples that must have \code{countMinimum} counts for an OTU to be retained.}

\item{nMinimumHaveCount}{numeric. If \code{filtering==TRUE}, the minimum number of samples that must have \code{countMinimum} counts for an OTU to be retained.}

\item{NA.RM}{logical. Whether or not to remove samples with NA values in the \code{sample_data()}.}

\item{...}{further arguments passed to \code{get_edgeR_results}}

\item{vars}{character vector, containing names of each independent variable for which you would like a separate model.}

\item{varlabels}{character vector, containing variable labels for printing}

\item{adjusted_for}{character vector, containing names of the additional adjustment variables to include in each model}

\item{to.data.frame}{logical. Whether or not to return a combined data.frame of all model results, rather than the default list.}
}
\description{
Returns an object of class \code{topTags} containing results filtered by FDR < alpha for one comparison. \code{get_edgeR_results_all_levels()} returns a list of \code{topTags} objects returned by \code{get_edgeR_results()} for all levels of the first variable in the formula.
}
