% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadQiimeData.R
\name{loadQiimeData}
\alias{loadQiimeData}
\title{Import NYC HANES-II microbiome data as a \code{phyloseq} object.}
\usage{

  loadQiimeData(metadata = sas7bdat::read.sas7bdat("http://nychanes.org/wp-content/uploads/sites/6/2018/01/public_v2_010518.sas7bdat"))
}
\description{
Import NYC HANES-II microbiome data as a \code{phyloseq} object.
}
