% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadQiimeData.R
\name{loadQiimeData}
\alias{loadQiimeData}
\title{Import}
\usage{
loadQiimeData()
}
\value{
A phyloseq object.
}
\description{
\code{loadQiimeData} imports NYC HANES-II microbiome data as a 
\code{phyloseq} object
}
\examples{

nychanes <- loadQiimeData()

}
